Reader: move post-key util to client/reader
