+++ client/state/ui/reader/card-expansions/test/reducer.js
@@ -10,7 +10,7 @@ import freeze from 'deep-freeze';
  * Internal dependencies
  */
 import reducer from '../reducer';
-import { keyToString } from 'lib/feed-stream-store/post-key';
+import { keyToString } from 'reader/post-key';
 import { READER_EXPAND_CARD, READER_RESET_CARD_EXPANSIONS } from 'state/action-types';
 
 describe( 'reducer', () => {
